cosmetics

+++ include_ps5/kernel.h
@@ -16,8 +16,8 @@ extern const off_t KERNEL_OFFSET_UCRED_CR_RUID;
 extern const off_t KERNEL_OFFSET_UCRED_CR_SVUID;
 extern const off_t KERNEL_OFFSET_UCRED_CR_RGID;
 
-int64_t kernel_get_fw_version(void);
-int32_t kernel_copyin(const void *udaddr, uint64_t kaddr, size_t len);
-int32_t kernel_copyout(const uint64_t kaddr, void *udaddr, size_t  len);
+uint32_t kernel_get_fw_version(void);
+int32_t  kernel_copyin(const void *udaddr, intptr_t kaddr, size_t len);
+int32_t  kernel_copyout(const intptr_t kaddr, void *udaddr, size_t  len);
 
 #endif // PS5SDK_KERNEL_H
